Move common cluster dict types to cluster_state
Move clusterNodesDictType, clusterSdsToListType, and clusterSlotDictType
from cluster_legacy.c to cluster_state.c and declare them in
cluster_state.h. These dict types back the common clusterState dicts
(nodes, shards, migrating/importing slots) and are needed by any cluster
bus implementation.

The legacy-specific clusterNodesBlackListDictType remains in
cluster_legacy.c.
